requirements.txt
streamlit==1.45.1
azure-storage-blob==12.25.1
azure-core==1.34.0
pymssql==2.3.4
python-dotenv==1.0.1
Pillow==10.3.0
